hyuck

+++ src/dungeon/floor_builder/floor_builder_state/has_connections.rs
@@ -47,7 +47,7 @@ impl FloorBuilder<HasConnections> {
                                 (
                                     p,
                                     if use_noise_map {
-                                        *self.noise_map.at(p, self.width)
+                                        *self.noise_map.at(p, self.width) as u32
                                     } else {
                                         1
                                     },
